Made Site Dynamic
First commit in a while, lots of changes to the site. This is difficult
to manage because I use Notepad++ as my FTP, but I've now mapped the
cache right into this folder. The site is now dynamic, and uses AJAX
instead of strict PHP. Next goal: add profile service