Portfolio page bugfix
The method for 'symbol' and 'price' extraction used the .sibling()
jQuery function to find each div. This would work fine unless the user
clicked on the 'symbol' or 'price' div. The new implementation uses a
different (albeit much more complicated) approach to extracting the
values needed for the portfolio delete query.